Change package import to relative (#19)
When a developer uses this template, they are likely to rename the
package from `voice_assistant` to something else for their app. This
absolute package-level import would break when they did so. I changed it
to relative, which we already do in the other spots.

Related, but not fixed, we do use a named package import in the sample
test but I'm not sure how to fix it.